bot rework

+++ bot.py
@@ -1,20 +1,17 @@
 #graczy w sensie gry nazywami botami
 from subprocess import Popen,PIPE,CREATE_NEW_CONSOLE
 import shlex
-from os import remove
+from server import Server
+from constants import host, port
+
+server = Server(host, port)
 
 class Bot:
     def __init__(self, name, prompt=None):
         self.prompt=prompt
         if prompt is None:
-            try:
-                with open(f'{name}.out', 'x') as f:
-                    pass
-            except FileExistsError:
-                pass
             self.proc=Popen(shlex.split(f"python terminalbot.py {name}"), creationflags=CREATE_NEW_CONSOLE)
-        else:
-            self.proc=Popen(shlex.split(prompt), stdin=PIPE, stdout=PIPE)
+
         self.name=name
         self.coins=0
         self.soliderLimit=8
@@ -22,6 +19,7 @@ class Bot:
         self.philosophers=0
         self.priests=0
         self.ownedTiles=[]
+
     def __hash__(self):
         return hash(self.name)
     def __eq__(self, other):
@@ -29,40 +27,16 @@ class Bot:
     def __ne__(self, other):
         return self.name!=other.name
     def get_move(self):
-        if self.prompt is not None:
-            return self.proc.stdout.readline().decode().split()
-        #for testing only
-        else:
-            print(f'ruch gracza {self.name}')
-            while True:
-                f=open(f'{self.name}.out', 'r')
-                d=f.readline().strip().split()
-                rest=f.read()
-                if d!=[]:
-                    print(d)
-                    f.close()
-                    with open(f'{self.name}.out', 'w') as f:
-                        f.write(rest)
-                    return d
-                f.close()
-    def send_move(self,player,move):
-        if self.prompt is not None:
-            return self.proc.stdin.write(f"{player} {" ".join(map(str, move))}")
-        #for testing only
-        else:
-            with open(f"{self.name}.in", "a") as f:
-                print(player, " ".join(map(str, move)), file=f)
+        print(f'ruch {self.name}')
+        while not server.data[self.name]:
+            pass
+        ret=server.data[self.name].pop(0).split()
+        print(f"ruch to {ret}")
+        return ret
+    
+    def send_move(self, player,move):
+        server.senddata(self.name, f'{player} {move}')
 
     def __del__(self):
-        self.proc.kill()
-        while self.proc.poll() is None:
-            pass
-        #if self.prompt is None:
-        #    remove(f'{self.name}.in')
-        #    remove(f'{self.name}.out')
-
-
-if __name__=='__main__':
-    b = Bot("x")
-    m=b.get_move()
-    print(m)
+        if hasattr(self, 'proc'):
+            self.proc.kill()
